Create component for radio menu option
