+++ src/components/editor/Bindings/FieldSelection/FieldActions/MenuOptions.tsx
@@ -1,17 +1,9 @@
-import {
-    Box,
-    FormControlLabel,
-    Radio,
-    RadioGroup,
-    Typography,
-} from '@mui/material';
-import { useIntl } from 'react-intl';
+import { RadioGroup } from '@mui/material';
 import { SelectionAlgorithm } from 'stores/Binding/slices/FieldSelection';
 import { useBindingStore } from 'stores/Binding/Store';
+import MenuOption from './MenuOption';
 
 export default function MenuOptions() {
-    const intl = useIntl();
-
     const selectionAlgorithm = useBindingStore(
         (state) => state.selectionAlgorithm
     );
@@ -27,70 +19,16 @@ export default function MenuOptions() {
             value={selectionAlgorithm}
             style={{ maxWidth: 320, textWrap: 'wrap' }}
         >
-            <FormControlLabel
+            <MenuOption
+                descriptionId="fieldSelection.massActionMenu.recommended.description"
+                labelId="fieldSelection.massActionMenu.recommended.label"
                 value="recommended"
-                control={<Radio size="small" />}
-                label={
-                    <Box style={{ padding: '8px 0px' }}>
-                        <Typography
-                            sx={{
-                                mb: '4px',
-                                fontWeight: 500,
-                            }}
-                        >
-                            {intl.formatMessage({
-                                id: 'fieldSelection.massActionMenu.recommended.label',
-                            })}
-                        </Typography>
-
-                        <Typography
-                            sx={{
-                                'textTransform': 'lowercase',
-                                '&.MuiTypography-root:first-letter': {
-                                    textTransform: 'uppercase',
-                                },
-                            }}
-                        >
-                            {intl.formatMessage({
-                                id: 'fieldSelection.massActionMenu.recommended.description',
-                            })}
-                        </Typography>
-                    </Box>
-                }
-                style={{ alignItems: 'flex-start' }}
             />
 
-            <FormControlLabel
+            <MenuOption
+                descriptionId="fieldSelection.massActionMenu.excludeAll.description"
+                labelId="fieldSelection.massActionMenu.excludeAll.label"
                 value="excludeAll"
-                control={<Radio size="small" />}
-                label={
-                    <Box style={{ padding: '8px 0px' }}>
-                        <Typography
-                            sx={{
-                                mb: '4px',
-                                fontWeight: 500,
-                            }}
-                        >
-                            {intl.formatMessage({
-                                id: 'fieldSelection.massActionMenu.excludeAll.label',
-                            })}
-                        </Typography>
-
-                        <Typography
-                            sx={{
-                                'textTransform': 'lowercase',
-                                '&.MuiTypography-root:first-letter': {
-                                    textTransform: 'uppercase',
-                                },
-                            }}
-                        >
-                            {intl.formatMessage({
-                                id: 'fieldSelection.massActionMenu.excludeAll.description',
-                            })}
-                        </Typography>
-                    </Box>
-                }
-                style={{ alignItems: 'flex-start' }}
             />
         </RadioGroup>
     );
